refactor: remove query.ts queryKey.ts files
